# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MyApp API",
    "version": "1.0.0",
    "description": "API documentation generated from Next.js API routes"
  },
  "paths": {
...
    "/api/games": {
      "post": {
        "summary": "Add one or multiple games to the database",
        "tags": [
          "Games"
        ],
        "requestBody": {
          "description": "Game object or array of game objects to insert",
          "required": true,
          "content": {
            "application/json": null
          }
        },
        "responses": {
          "200": {
            "description": "Games inserted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "insertedCount": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, missing games or hash field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}